identity: execute method and add key using txManager (#729)
* added testworld tests for proofs

* bind identity service to did

* removed identity init from bootstrapper

* added account context to create identity

* added waitingTransaction task

* create identity uses task manager

* refactored add key and execute tests

* correct usage of done flag

* added context to identity methods

* formatting

* fixed add key testcase

* add key uses txManager

* using tx manager for exceute methods

* removed old wait for tx helper method

* fixed correct reset of identityID

* refactored identity to service

* refactored service to factory

* refactored identity to service

* removed config from factory

* small formatting issues
